Show the latest idle time for the idle in TX.
When the idle in transaction backend is running, it might
execute multiple queries sequentially and still marked as
idle in transaction during the pauses between queries. In
order to distinguish betwen processs that are actively doing
something and those that are left idle for hours we show the
most recent idle interval for the idle in transaction backend.

We don't show any additional information if transaction age
is the same as the length of the idle time to avoid cluttering
the output.

For 9.1 and below this are more complicated, since there is no
way to discover when the backend state has transitioned to the
idle in transaction, so we show the time since the last query
start.